rename benchmarks to benchmark
