+++ tests/cog/agent/test_self_refine.py
@@ -11,7 +11,7 @@ from agential.cog.prompts.agents.self_refine import (
     SELF_REFINE_INSTRUCTION_GSM8K,
     SELF_REFINE_REFINE_INSTRUCTION_GSM8K,
 )
-from agential.cog.prompts.benchmarks.gsm8k import GSM8K_FEWSHOT_EXAMPLES_POT
+from agential.cog.prompts.benchmark.gsm8k import GSM8K_FEWSHOT_EXAMPLES_POT
 from agential.cog.strategies.self_refine.base import SelfRefineBaseStrategy
