+++ tests/cog/functional/test_reflexion_functional.py
@@ -34,7 +34,7 @@ from agential.cog.prompts.agents.reflexion import (
     REFLEXION_COT_REFLECT_FEWSHOT_EXAMPLES_NO_CONTEXT,
     REFLEXION_REACT_REFLECT_FEWSHOT_EXAMPLES,
 )
-from agential.cog.prompts.benchmarks.hotpotqa import HOTPOTQA_FEWSHOT_EXAMPLES_REACT
+from agential.cog.prompts.benchmark.hotpotqa import HOTPOTQA_FEWSHOT_EXAMPLES_REACT
 
 
 def test__truncate_scratchpad() -> None:
