+++ tests/cog/strategies/self_refine/test_math.py
@@ -9,7 +9,7 @@ from agential.cog.prompts.agents.self_refine import (
     SELF_REFINE_INSTRUCTION_GSM8K,
     SELF_REFINE_REFINE_INSTRUCTION_GSM8K,
 )
-from agential.cog.prompts.benchmarks.gsm8k import GSM8K_FEWSHOT_EXAMPLES_POT
+from agential.cog.prompts.benchmark.gsm8k import GSM8K_FEWSHOT_EXAMPLES_POT
 from agential.cog.strategies.self_refine.math import (
     SelfRefineGSM8KStrategy,
     SelfRefineMathStrategy,
